changed our_number column to be a bigint

+++ segue/purchase/boleto/models.py
@@ -14,7 +14,7 @@ class Weekdays():
 class BoletoPayment(Payment):
     __mapper_args__ = { 'polymorphic_identity': 'boleto' }
 
-    our_number    = db.Column(db.Integer,    name='bo_our_number')
+    our_number    = db.Column(db.BigInteger, name='bo_our_number')
     due_date      = db.Column(db.Date,       name='bo_due_date')
     document_hash = db.Column(db.String(32), name='bo_document_hash')
